sample-application/requirements.txt
langchain>=0.0.267
openai==1.13.3
streamlit
tiktoken==0.3.3
llama-index==0.10.15
llama-index-llms-langchain==0.1.3
llama-index-embeddings-langchain==0.1.2
python-dotenv==1.0.0
pypdf==4.1.0
azure-core==1.26.4
azure-identity==1.13.0
hvac==0.10.4
